Wave 12.5: Fix unified Plaid sync bugs and loan analytics
Plaid sync fixes:
- Fix IsCreditCard LINQ translation error (inline type check vs unmapped property)
- Fix DateTime Kind=Unspecified errors for Npgsql timestamptz columns (3 locations)
- Fix credential encryption/decryption in PlaidLiabilitiesService
- Fix cascade deletion: DeleteConnectionAsync now removes liabilities, CC transactions, and linked properties
- Fix DisconnectAsync to mark liabilities as disconnected

Loan analytics fixes:
- Fix negative amortization handling when min payment < monthly interest
- Fix percentPaidOff: use actual balance ratio instead of schedule-based principal
- Clamp progress bar and chart values to 0-100%
- Add warning indicators for income-driven/deferred interest plans
- Add ActualMinimumPayment field to DTO and frontend types
- Change label from 'remaining' to 'original' in loan summary

Other:
- Add PlaidConnectionService encryption for unified token exchange
- Add PlaidInvestmentsService encryption support
- Add PriceRefreshJob improvements
- Add Wave 12.5 testing guide and custom Plaid sandbox user configs
- Update documentation map

+++ PFMP-API/Services/MarketData/IMarketDataService.cs
@@ -24,8 +24,8 @@ public class FmpQuote
     public decimal? YearHigh { get; set; }
     public decimal? YearLow { get; set; }
     public decimal? MarketCap { get; set; }
-    public long? Volume { get; set; }
-    public long? AvgVolume { get; set; }
+    public decimal? Volume { get; set; }
+    public decimal? AvgVolume { get; set; }
     public string Exchange { get; set; } = string.Empty;
     public decimal? Open { get; set; }
     public decimal? PreviousClose { get; set; }
